+++ config.js
@@ -4,5 +4,6 @@ export const settings = {
     "default_time_zone": "Central",
     "default_meridiem": "PM",
     "default_poll_options": 5,
-    "default_calendar_channel": "calendar"
+    "default_calendar_channel": "calendar",
+    "default_poll_channel": "scheduling"
 }